API Reference

DispatcherCache.__hash

DispatcherCache.get_compressor

DispatcherCache.get_hash

DispatcherCache.input_nodes

DispatcherCache.load_hashchain

DispatcherCache.store_hashchain

DispatcherCache.__hash — Method.

__hash(something)

Return a hexadecimal string corresponding to the hash of sum of the hashes of the value and type of something.

source

DispatcherCache.get_compressor — Method.

get_compressor(compression, action)

Return a TranscodingStreams compatible compressor or decompressor based on the values of compression and action.

source

DispatcherCache.get_hash — Method.

get_hash(node, keyhashmaps)

Calculates and returns the hash corresponding to a Dispatcher task graph node i.e. DispatchNode using the hashes of its dependencies, input arguments and source code of the function associated to the node. Any available hashes are taken from keyhashmap.

source

DispatcherCache.input_nodes — Method.

input_nodes(graph::DispatchGraph) ->

Return an iterable of all nodes in the graph with no input edges.

source

DispatcherCache.load_hashchain — Function.

load_hashchain(cachedir [; compression=DEFAULT_COMPRESSION])

Loads the hashchain file found in the directory cachedir. Before loading, the compression value is checked against the one stored in the hashchain file (both have to match). If the file does not exist, it is created.

source

DispatcherCache.store_hashchain — Function.

store_hashchain(hashchain, cachedir=DEFAULT_CACHE_DIR [; compression=DEFAULT_COMPRESSION, version=1])

Stores the hashchain object in a file named DEFAULT_HASHCHAIN_FILENAME, in the directory cachedir. The values of compression and version are stored as well in the file.

source