API Reference

DispatcherCache.DispatcherCache

DispatcherCache.__hash

DispatcherCache.cache!

DispatcherCache.get_arguments_hash

DispatcherCache.get_compressor

DispatcherCache.get_dependencies_hash

DispatcherCache.get_hash

DispatcherCache.get_source_hash

DispatcherCache.input_nodes

DispatcherCache.load_hashchain

DispatcherCache.store_hashchain

DispatcherCache.DispatcherCache — Module.

DispatcherCache.jl is a hash-chain optimizer for Dispatcher delayed execution graphs. It employes a hashing mechanism to check wether the state associated to a node is the DispatchGraph that is to be executed has already been hashed (and hence, an output is available) or, it is new or changed. Depending on the current state (by 'state' one understands the called function source code, input arguments and other input node dependencies), the current task becomes a load-from-disk or an execute-and-store-to-disk operation. This is done is such a manner that the minimimum number of load/execute operations are performed, minimizing both persistency and computational demands.

source

DispatcherCache.cache! — Method.

cache!(graph, endpoints, uncacheable, compression=DEFAULT_COMPRESSION, cachedir=DEFAULT_CACHE_DIR)

Optimizes a delayed execution graph graph::DispatchGraph by wrapping individual nodes in load-from-disk on execute-and-store wrappers depending on the state of the disk cache and of the graph. The function modifies inplace the input graph and should be called on the same unmodified graph after each execution and not on the modified graph.

Once the original graph is modified, calling run! on it will, if the cache is already present, load the top most consistent key or alternatively re-run and store the outputs of nodes which have new state.

Arguments

graph::DispatchGraph input graph

endpoints::Vector{Union{DispatchNode, AbstractString}} top nodes for which

caching will occur; nodes that depend on these will not be cached. The nodes can be specified either by label of by the node object itself

uncacheable::Vector{Union{DispatchNode, AbstractString}}` nodes that will

never be cached and will always be executed (these nodes are still hashed and their hashes influence upstream node hashes as well)

Keyword arguments

compression::String enables compression of the node outputs.

Available options are "none", for no compression, "bz2" or "bzip2" for BZIP compression and "gz" or "gzip" for GZIP compression

cachedir::String The cache directory.

Examples

julia> # TODO ...

source

DispatcherCache.__hash — Method.

__hash(something)

Return a hexadecimal string corresponding to the hash of sum of the hashes of the value and type of something.

source

DispatcherCache.get_arguments_hash — Method.

get_arguments_hash(node)

Hash the data arguments (in certain cases configuration fields) of the dispatch node.

source

DispatcherCache.get_compressor — Method.

get_compressor(compression, action)

Return a TranscodingStreams compatible compressor or decompressor based on the values of compression and action.

source

DispatcherCache.get_dependencies_hash — Method.

get_dependencies_hash(node, keyhashmaps)

Hash the dispatch node dependencies of node using their existing hashes if possible.

source

DispatcherCache.get_hash — Method.

get_hash(node, keyhashmaps)

Calculates and returns the hash corresponding to a Dispatcher task graph node i.e. DispatchNode using the hashes of its dependencies, input arguments and source code of the function associated to the node. Any available hashes are taken from keyhashmap.

source

DispatcherCache.get_source_hash — Method.

get_source_hash(node)

Hashes the lowered representation of the source code of the function associated with node. Useful for Op nodes, the other node types do not have any associated source code.

source

DispatcherCache.input_nodes — Method.

input_nodes(graph::DispatchGraph) ->

Return an iterable of all nodes in the graph with no input edges.

source

DispatcherCache.load_hashchain — Function.

load_hashchain(cachedir [; compression=DEFAULT_COMPRESSION])

Loads the hashchain file found in the directory cachedir. Before loading, the compression value is checked against the one stored in the hashchain file (both have to match). If the file does not exist, it is created.

source

DispatcherCache.store_hashchain — Function.

store_hashchain(hashchain, cachedir=DEFAULT_CACHE_DIR [; compression=DEFAULT_COMPRESSION, version=1])

Stores the hashchain object in a file named DEFAULT_HASHCHAIN_FILENAME, in the directory cachedir. The values of compression and version are stored as well in the file.

source